Standardize model-visible fragment registration
