+++ codex-rs/core/src/agent/control.rs
@@ -6,9 +6,9 @@ use crate::agent::status::is_final;
 use crate::error::CodexErr;
 use crate::error::Result as CodexResult;
 use crate::find_thread_path_by_id_str;
+use crate::model_visible_fragments::SubagentNotification;
+use crate::model_visible_fragments::format_subagent_context_line;
 use crate::rollout::RolloutRecorder;
-use crate::session_prefix::SubagentNotification;
-use crate::session_prefix::format_subagent_context_line;
 use crate::shell_snapshot::ShellSnapshot;
 use crate::state_db;
 use crate::thread_manager::ThreadManagerState;
