+++ codex-rs/core/src/event_mapping.rs
@@ -18,7 +18,7 @@ use codex_protocol::user_input::UserInput;
 use tracing::warn;
 use uuid::Uuid;
 
-use crate::model_visible_context::is_contextual_user_fragment;
+use crate::model_visible_fragments::is_contextual_user_fragment;
 use crate::web_search::web_search_action_detail;
 
 pub(crate) fn is_contextual_user_message_content(message: &[ContentItem]) -> bool {
