--- codex-rs/core/src/instructions/mod.rs
@@ -1,5 +1,0 @@
-mod contextual_user_fragments;
-
-pub(crate) use contextual_user_fragments::AgentsMdInstructions;
-pub(crate) use contextual_user_fragments::PluginInstructions;
-pub(crate) use contextual_user_fragments::SkillInstructions;
