+++ codex-rs/core/src/lib.rs
@@ -31,7 +31,6 @@ pub mod connectors;
 mod context_manager;
 pub mod custom_prompts;
 pub mod env;
-mod environment_context;
 pub mod error;
 pub mod exec;
 pub mod exec_env;
@@ -42,12 +41,12 @@ mod file_watcher;
 mod flags;
 pub mod git_info;
 mod guardian;
-pub mod instructions;
 pub mod landlock;
 pub mod mcp;
 mod mcp_connection_manager;
 mod mcp_tool_approval_templates;
 mod model_visible_context;
+mod model_visible_fragments;
 pub mod models_manager;
 mod network_policy_decision;
 pub mod network_proxy_loader;
@@ -67,7 +66,6 @@ pub mod personality_migration;
 pub mod plugins;
 mod sandbox_tags;
 pub mod sandboxing;
-mod session_prefix;
 mod shell_detect;
 mod stream_events_utils;
 pub mod test_support;
@@ -145,7 +143,6 @@ pub use rollout::session_index::find_thread_names_by_ids;
 mod function_tool;
 mod state;
 mod tasks;
-mod user_shell_command;
 pub mod util;
 pub(crate) use codex_protocol::protocol;
 pub(crate) use codex_shell_command::bash;
