+++ codex-rs/core/src/skills/injection.rs
@@ -6,10 +6,10 @@ use crate::analytics_client::AnalyticsEventsClient;
 use crate::analytics_client::InvocationType;
 use crate::analytics_client::SkillInvocation;
 use crate::analytics_client::TrackEventsContext;
-use crate::instructions::SkillInstructions;
 use crate::mention_syntax::TOOL_MENTION_SIGIL;
 use crate::mentions::build_skill_name_counts;
 use crate::model_visible_context::ModelVisibleContextFragment;
+use crate::model_visible_fragments::SkillInstructions;
 use crate::skills::SkillMetadata;
 use codex_otel::SessionTelemetry;
 use codex_protocol::models::ResponseItem;
